Profile: Add pssn visibility toggle and consent update (#114)
* Add pssn toggle and consent update

* Update user details migration

* Fix date of birth type
