+++ apps/profile/app/components/AboutMe.tsx
@@ -4,115 +4,6 @@ import UserDetails from "./UserDetails";
 import Addresses from "./Addresses";
 import Entitlements from "./Entitlements";
 import Consent from "./Consent";
-import { revalidatePath } from "next/cache";
-import { form, postgres } from "../utils";
-import { PgSessions } from "auth/sessions";
-
-async function submitAction(formData: FormData) {
-  "use server";
-
-  const { userId } = await PgSessions.get();
-
-  const formErrors: form.Error[] = [];
-
-  const phone = formData.get("phone")?.toString();
-  formErrors.push(
-    ...form.validation.stringNotEmpty(form.fieldTranslationKeys.phone, phone),
-  );
-
-  const email = formData.get("email")?.toString();
-  formErrors.push(
-    ...form.validation.emailErrors(form.fieldTranslationKeys.email, email),
-  );
-
-  if (formErrors.length) {
-    await form.insertErrors(formErrors, userId);
-
-    return revalidatePath("/");
-  }
-
-  const dayOfBirth = formData.get("dayOfBirth");
-  const monthOfBirth = formData.get("monthOfBirth");
-  const yearOfBirth = formData.get("yearOfBirth");
-
-  const dateOfBirth = new Date(
-    Number(yearOfBirth),
-    Number(monthOfBirth) - 1,
-    Number(dayOfBirth),
-  );
-
-  let data = {
-    user_id: userId,
-    date_of_birth: dateOfBirth,
-    title: "",
-    firstName: "",
-    lastName: "",
-    ppsn: "",
-    gender: "",
-    phone: "",
-    email: "",
-  };
-
-  const formIterator = formData.entries();
-  let iterResult = formIterator.next();
-
-  while (!iterResult.done) {
-    const [key, value] = iterResult.value;
-
-    if (
-      [
-        "title",
-        "firstName",
-        "lastName",
-        "ppsn",
-        "gender",
-        "phone",
-        "email",
-      ].includes(key)
-    ) {
-      data[key] = value;
-    }
-
-    iterResult = formIterator.next();
-  }
-
-  const currentDataResults = await postgres.pgpool.query(
-    `
-      SELECT * FROM user_details
-      WHERE user_id = $1
-  `,
-    [userId],
-  );
-
-  const keys = Object.keys(data);
-  const values = Object.values(data);
-
-  if (currentDataResults.rows.length > 0) {
-    const setClause = keys
-      .map((key, index) => `${key} = $${index + 1}`)
-      .join(", ");
-
-    await postgres.pgpool.query(
-      `
-        UPDATE user_details
-        SET ${setClause}
-        WHERE user_id = $${keys.length + 1}
-      `,
-      [...values, userId],
-    );
-  } else {
-    const columns = keys.join(", ");
-    const placeholders = values.map((_, index) => `$${index + 1}`).join(", ");
-
-    await postgres.pgpool.query(
-      `
-        INSERT INTO user_details (${columns})
-        VALUES (${placeholders})
-      `,
-      values,
-    );
-  }
-}
 
 export default async () => {
   const t = await getTranslations("AboutMe");
@@ -146,34 +37,13 @@ export default async () => {
           })}
         </strong>
       </p>
-      <form action={submitAction}>
-        <UserDetails />
-        <hr style={{ marginBottom: "30px" }} />
-        <Addresses />
-        <hr style={{ marginBottom: "30px" }} />
-        <Entitlements />
-        <hr style={{ marginBottom: "30px" }} />
-        <Consent />
-        <hr style={{ marginBottom: "30px" }} />
-        <div style={{ display: "flex", gap: "20px", alignItems: "center" }}>
-          <button
-            type="submit"
-            data-module="govie-button"
-            className="govie-button"
-            style={{ marginBottom: 0 }}
-          >
-            {t("save")}
-          </button>
-          <button
-            type="button"
-            data-module="govie-button"
-            className="govie-button govie-button--secondary"
-            style={{ marginBottom: 0 }}
-          >
-            {t("cancel")}
-          </button>
-        </div>
-      </form>
+      <UserDetails />
+      <hr style={{ marginBottom: "30px" }} />
+      <Addresses />
+      <hr style={{ marginBottom: "30px" }} />
+      <Entitlements />
+      <hr style={{ marginBottom: "30px" }} />
+      <Consent />
     </div>
   );
 };
