+++ apps/profile/package.json
@@ -6,7 +6,8 @@
     "build": "next build",
     "start": "next start -p 3003",
     "test:e2e": "playwright test",
-    "migrate": "db-migrate up"
+    "migrate": "db-migrate up",
+    "migrate:create": "db-migrate create $npm_config_name --sql-file"
   },
   "nodemonConfig": {
     "watch": [
